Moved color picker palette to common
Refactored the ColorPickerPalette container out to be its own component so different triggers can use the same palette
