+++ v3/src/components/common/color-picker.tsx
