+++ v3/src/components/common/point-color-setting-shared.scss
