+++ v3/src/components/common/point-color-setting.scss
@@ -18,8 +18,6 @@
     border: 1px solid #d0d0d0;
     top: shared.$color-picker-popover-top-px;
     left: shared.$color-picker-popover-left-px;
-    // top: -8px;
-    // left: 100px;
 
     .color-swatch-palette {
       display: flex;
